Ticket 3 (Marcus Schmid): Datentypen harmonisiert, Initialisierungen in potentiell ungefährlichere Reihenfolge gebracht, unnötige Variablen entfernt, Formatstrings angepasst
Letzten Endes: den Compiler-Hinweisen bei -Wall -pedantic nachgegangen; hoffentlich nix dabei kaputt gemacht ;)

+++ trunk/Beryllium10/Code/Interfaces/IProvider.cpp
@@ -145,7 +145,7 @@ void IProvider::GetAllTextFromNode( wxXmlNode *node, wxString &text )
 // Hilfsfunktion: Ersetzt alle Vorkommen von "findstr" mit "replacewith"
 void IProvider::ReplaceAll( wxString &text, wxString findstr, wxString replacewith )
 {
-	int pos = text.find(findstr);
+	size_t pos = text.find(findstr);
 
 	while( pos != wxString::npos)
 	{
